IDENT-4593: MARKETING_VERSION from pre processor macro (#314)
In order to bypass Apples version management solution which sets all frameworks version to products MARKETING_VERSION, a pre processore macro has been introduced both for project and cocoapods configuration. A simple objc definition gets the preprocessor value for MARKETING_VERSION and populates swift based static variable that can be easily read from swift code.

The only caveat for this solution is that CoreVersion.marketingVersion becomes public variable as this is necessary for the Objective C extension to be defined in the first place.